// src/models/Vehicle/VehicleSpecifications.ts
export interface VehicleSpecifications {
  make: string;
  type?: string;
  subType?: string;
  model: string;
  year?: number;
  color?: string;
  mileage?: number;
  fuelType?: string;
  transmission?: string;
  vin?: string;
  cylinders?: number;
  displacement?: number;
  fuelEfficiency?: number;
}